Добавляет компонент main-empty
