+++ src/pages/main/main.tsx
@@ -1,30 +1,17 @@
-import { useState } from 'react';
+import cn from 'classnames';
 
 import HeaderNav from '../../components/header-nav/header-nav';
 import Header from '../../components/header/header';
 import Logo from '../../components/logo/logo';
-import OffersList from '../../components/offers-list/offers-list';
-import Map from '../../components/map/map';
 
 import { useAppSelector } from '../../hooks';
-import { TOffer } from '../../types/offer';
 import TabsList from '../../components/tabs-list/tabs-list';
+import MainContent from '../../components/main-content/main-content';
+import MainEmpty from '../../components/main-empty/main-empty';
 
 function Main(): JSX.Element {
-  const [hoveredOffer, setHoveredOffer] = useState<TOffer | undefined>();
-
   const city = useAppSelector((state) => state.city);
   const offers = useAppSelector((state) => state.offers);
-
-  function handleCardHover(id: number) {
-    const currentOffer = offers.find((offer) => id === offer.id);
-
-    if (!currentOffer) {
-      return;
-    }
-
-    setHoveredOffer(currentOffer);
-  }
 
   return (
     <div className="page page--gray page--main">
@@ -34,7 +21,11 @@ function Main(): JSX.Element {
           <HeaderNav />
         </>
       </Header>
-      <main className="page__main page__main--index">
+      <main
+        className={cn('page__main page__main--index', {
+          'page__main--index-empty': offers.length === 0
+        })}
+      >
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
           <section className="locations container">
@@ -42,38 +33,11 @@ function Main(): JSX.Element {
           </section>
         </div>
         <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{offers.length} places to stay in {city}</b>
-              <form className="places__sorting" action="#" method="get">
-                <span className="places__sorting-caption">Sort by </span>
-                <span className="places__sorting-type" tabIndex={0}>
-                  Popular
-                  <svg className="places__sorting-arrow" width="7" height="4">
-                    <use xlinkHref="#icon-arrow-select"></use>
-                  </svg>
-                </span>
-                {/* places__options--opened */}
-                <ul className="places__options places__options--custom">
-                  <li className="places__option places__option--active" tabIndex={0}>Popular</li>
-                  <li className="places__option" tabIndex={0}>Price: low to high</li>
-                  <li className="places__option" tabIndex={0}>Price: high to low</li>
-                  <li className="places__option" tabIndex={0}>Top rated first</li>
-                </ul>
-              </form>
-              <OffersList
-                offers={offers}
-                onCardHover={handleCardHover}
-              />
-            </section>
-            <div className="cities__right-section">
-              <Map
-                offers={offers}
-                hoveredOffer={hoveredOffer}
-              />
-            </div>
-          </div>
+          {
+            offers.length !== 0
+              ? <MainContent city={city} offers={offers} />
+              : <MainEmpty city={city} />
+          }
         </div>
       </main>
     </div>
