+++ src/store/reducer.ts
@@ -19,6 +19,10 @@ const reducer = createReducer(initialState, (builder) => {
   builder.addCase(changeCity, (state, action) => {
     const {city} = action.payload;
 
+    if (city === state.city) {
+      return;
+    }
+
     state.city = city;
     state.offers = offers.filter((offer) => offer.city.name === city);
   });
